# frontend/smart-note-frontend/api-docs.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8801","description":"Generated server url"}],"paths":{"/api/users/password":{"put":{"tags":["user-controller"],"operationId":"updatePassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePasswordDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}}},"/api/users/motto":{"put":{"tags":["user-controller"],"operationId":"updateMotto","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateMottoDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}}},"/api/users/avatar":{"put":{"tags":["user-controller"],"operationId":"updateAvatar","requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}},"required":["file"]}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultString"}}}}}}},"/api/notes/{id}":{"get":{"tags":["note-controller"],"operationId":"getNote","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultNoteDetailVO"}}}}}},"put":{"tags":["note-controller"],"operationId":"updateNote","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NoteUpdateDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}},"delete":{"tags":["note-controller"],"operationId":"deleteNote","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}}},"/api/notes/{id}/visibility":{"put":{"tags":["note-controller"],"operationId":"setVisibility","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"shareLevel","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}}},"/api/notes/{id}/trash":{"put":{"tags":["note-controller"],"operationId":"trashNote","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}}},"/api/friends/{friendId}/group":{"put":{"tags":["friend-controller"],"operationId":"updateFriendGroup","parameters":[{"name":"friendId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FriendGroupDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}}},"/api/friends/requests/{requestId}/reject":{"put":{"tags":["friend-controller"],"operationId":"rejectRequest","parameters":[{"name":"requestId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}}},"/api/friends/requests/{requestId}/accept":{"put":{"tags":["friend-controller"],"operationId":"acceptRequest","parameters":[{"name":"requestId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}}},"/api/users/register":{"post":{"tags":["user-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}}},"/api/users/login":{"post":{"tags":["user-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultLoginVO"}}}}}}},"/api/notes":{"get":{"tags":["note-controller"],"operationId":"listNotes","parameters":[{"name":"pageNum","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":1}},{"name":"pageSize","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"title","in":"query","required":false,"schema":{"type":"string"}},{"name":"tag","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultPageResultNote"}}}}}},"post":{"tags":["note-controller"],"operationId":"createNote","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NoteCreateDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}}},"/api/notes/{id}/analyze":{"post":{"tags":["note-controller"],"operationId":"analyzeNote","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultAIAnalysisVO"}}}}}}},"/api/friends/requests/{friendId}":{"post":{"tags":["friend-controller"],"operationId":"sendRequest","parameters":[{"name":"friendId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"groupTag","in":"query","required":false,"schema":{"type":"string","default":"默认"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}}},"/api/users/search":{"get":{"tags":["user-controller"],"operationId":"searchUsers","parameters":[{"name":"keyword","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultListUser"}}}}}}},"/api/notes/share/{noteId}":{"get":{"tags":["note-controller"],"operationId":"shareNote","parameters":[{"name":"noteId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultNoteDetailVO"}}}}}}},"/api/notes/history":{"get":{"tags":["note-controller"],"operationId":"getHistory","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultListNote"}}}}}}},"/api/friends":{"get":{"tags":["friend-controller"],"operationId":"listFriends","parameters":[{"name":"group","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultListUser"}}}}}}},"/api/friends/search":{"get":{"tags":["friend-controller"],"operationId":"search","parameters":[{"name":"keyword","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultListUser"}}}}}}},"/api/friends/requests":{"get":{"tags":["friend-controller"],"operationId":"getRequests","parameters":[{"name":"state","in":"query","required":false,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultListFriendship"}}}}}}},"/api/friends/groups":{"get":{"tags":["friend-controller"],"operationId":"getUserGroups","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultListString"}}}}}}},"/api/friends/{friendId}":{"delete":{"tags":["friend-controller"],"operationId":"deleteFriend","parameters":[{"name":"friendId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}}}},"components":{"schemas":{"UpdatePasswordDTO":{"type":"object","properties":{"oldPassword":{"type":"string"},"newPassword":{"type":"string","maxLength":20,"minLength":6}},"required":["newPassword","oldPassword"]},"ResultVoid":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"msg":{"type":"string"},"data":{"type":"object"}}},"UpdateMottoDTO":{"type":"object","properties":{"motto":{"type":"string"}}},"ResultString":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"msg":{"type":"string"},"data":{"type":"string"}}},"NoteUpdateDTO":{"type":"object","properties":{"title":{"type":"string"},"content":{"type":"string"},"tags":{"type":"string"}}},"FriendGroupDTO":{"type":"object","properties":{"groupTag":{"type":"string"}},"required":["groupTag"]},"RegisterDTO":{"type":"object","properties":{"account":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"password":{"type":"string","maxLength":20,"minLength":6},"confirmPassword":{"type":"string"}},"required":["account","confirmPassword","email","password","phone"]},"LoginDTO":{"type":"object","properties":{"keyword":{"type":"string"},"password":{"type":"string"}},"required":["keyword","password"]},"LoginVO":{"type":"object","properties":{"token":{"type":"string"},"userId":{"type":"integer","format":"int64"},"account":{"type":"string"},"username":{"type":"string"},"avatar":{"type":"string"}},"required":["token","userId"]},"ResultLoginVO":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"msg":{"type":"string"},"data":{"$ref":"#/components/schemas/LoginVO"}}},"NoteCreateDTO":{"type":"object","properties":{"title":{"type":"string"},"content":{"type":"string"},"tags":{"type":"string"}},"required":["content","title"]},"AIAnalysisVO":{"type":"object","properties":{"summary":{"type":"string"},"keyPoints":{"type":"string"},"tags":{"type":"string"}}},"ResultAIAnalysisVO":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"msg":{"type":"string"},"data":{"$ref":"#/components/schemas/AIAnalysisVO"}}},"ResultListUser":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"msg":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}},"User":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"account":{"type":"string"},"password":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"motto":{"type":"string"},"avatar":{"type":"string"},"createTime":{"type":"string","format":"date-time"},"updateTime":{"type":"string","format":"date-time"}}},"Note":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"userId":{"type":"integer","format":"int64"},"title":{"type":"string"},"content":{"type":"string"},"tags":{"type":"string"},"visibility":{"type":"string","enum":["PRIVATE","FRIEND_READ","FRIEND_WRITE","PUBLIC"]},"isDelete":{"type":"integer","format":"int32"},"createTime":{"type":"string","format":"date-time"},"updateTime":{"type":"string","format":"date-time"}}},"PageResultNote":{"type":"object","properties":{"total":{"type":"integer","format":"int64"},"pageNum":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"data":{"type":"array","items":{"$ref":"#/components/schemas/Note"}}}},"ResultPageResultNote":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"msg":{"type":"string"},"data":{"$ref":"#/components/schemas/PageResultNote"}}},"NoteDetailVO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"content":{"type":"string"},"tags":{"type":"string"},"shareLevel":{"type":"integer","format":"int32"},"summary":{"type":"string"},"keyPoints":{"type":"string"},"tagsSuggestion":{"type":"string"}}},"ResultNoteDetailVO":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"msg":{"type":"string"},"data":{"$ref":"#/components/schemas/NoteDetailVO"}}},"ResultListNote":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"msg":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/Note"}}}},"Friendship":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"userId":{"type":"integer","format":"int64"},"friendId":{"type":"integer","format":"int64"},"status":{"type":"string","enum":["WAITTING","ACCEPTED","REJECTED","DELETED"]},"groupTag":{"type":"string"},"creatTime":{"type":"string","format":"date-time"}}},"ResultListFriendship":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"msg":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/Friendship"}}}},"ResultListString":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"msg":{"type":"string"},"data":{"type":"array","items":{"type":"string"}}}}}}}